+++ packages/editor/editor-common/src/schema/create-schema.ts
@@ -204,7 +204,7 @@ export function createSchema(config: SchemaConfig): Schema {
   let nodes = addItems(nodesInOrder, nodesConfig, customNodeSpecs);
   let marks = addItems(marksInOrder, marksConfig, customMarkSpecs);
   nodes = sanitizeNodes(nodes, marks);
-  return new Schema({
+  return new Schema<string, string>({
     nodes,
     marks,
   });