+++ packages/editor/editor-core/src/plugins/image-upload/pm-plugins/commands.ts
@@ -18,9 +18,11 @@ export const insertExternalImage: (
     return false;
   }
 
-  dispatch(
-    safeInsert(mediaNode, state.selection.$to.pos)(state.tr).scrollIntoView(),
-  );
+  if (dispatch) {
+    dispatch(
+      safeInsert(mediaNode, state.selection.$to.pos)(state.tr).scrollIntoView(),
+    );
+  }
   return true;
 };
 
@@ -33,6 +35,8 @@ export const startImageUpload: (event?: Event) => Command = event => (
     return false;
   }
 
-  dispatch(startUpload(event)(state.tr));
+  if (dispatch) {
+    dispatch(startUpload(event)(state.tr));
+  }
   return true;
 };