+++ packages/editor/editor-core/src/ui/ContentStyles/index.ts
@@ -25,7 +25,9 @@ import { tasksAndDecisionsStyles } from '../../plugins/tasks-and-decisions/ui/st
 import { gridStyles } from '../../plugins/grid/styles';
 import { linkStyles } from '../../plugins/hyperlink/styles';
 
-const ContentStyles: ComponentClass<HTMLAttributes<{}>> = styled.div`
+const ContentStyles: ComponentClass<
+  HTMLAttributes<{}> & { theme: any }
+> = styled.div`
   /* Hack for ie11 that is being used in code block.
    * https://bitbucket.org/atlassian/atlaskit/src/ad09f6361109ece1aab316c8cbd8116ffb7963ef/packages/editor-core/src/schema/nodes/code-block.ts?fileviewer=file-view-default#code-block.ts-110
    */