+++ packages/elements/status/src/__tests__/unit/components/internal/colorSpec.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { mount } from 'enzyme';
 import EditorDoneIcon from '@atlaskit/icon/glyph/editor/done';
-import Color from '../../../components/internal/color';
+import Color from '../../../../components/internal/color';
+import { ANALYTICS_HOVER_DELAY } from '../../../../components/constants';
 
 describe('Color', () => {
   it('should render color button', () => {
@@ -64,5 +65,45 @@ describe('Color', () => {
 
     component.find('button').simulate('click');
     expect(onClick).toHaveBeenCalledWith(value);
+  });
+
+  describe('Color onHover', () => {
+    let realDateNow;
+    let dateNowStub;
+
+    beforeEach(() => {
+      realDateNow = Date.now;
+      dateNowStub = jest.fn();
+      Date.now = dateNowStub;
+    });
+
+    afterEach(() => {
+      Date.now = realDateNow;
+    });
+
+    it('should call onHover handler', () => {
+      const now = realDateNow();
+      const onHover = jest.fn();
+      const value = 'purple';
+      const component = mount(
+        <Color
+          value={value}
+          label={'Purple'}
+          onClick={jest.fn()}
+          onHover={onHover}
+          backgroundColor={'backgroundColor'}
+          borderColor={'borderColor'}
+          isSelected={false}
+        />,
+      );
+
+      dateNowStub.mockReturnValue(now);
+      component.find('button').simulate('mouseenter');
+
+      dateNowStub.mockReturnValue(now + ANALYTICS_HOVER_DELAY + 10);
+      component.find('button').simulate('mouseleave');
+
+      expect(onHover).toHaveBeenCalledWith(value);
+    });
   });
 });