+++ packages/editor/editor-common/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/editor-common
 
+## 22.4.0
+- [minor] [e06b553](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e06b553):
+
+  - ED-5702: default new table resizing
+
 ## 22.3.2
 - [patch] [60a4609](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/60a4609):