+++ packages/elements/mention/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/mention
 
+## 15.1.9
+- [patch] [b81da9b](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/b81da9b):
+
+  - Fix typescript types to support strictFunctionTypes
+
 ## 15.1.8
 - Updated dependencies [58b84fa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/58b84fa):
   - @atlaskit/analytics@4.0.7